move player filters + pagination to PlayerList component
